==> iris-classification/notebooks/02-train-pipeline.ipynb <==
# Train pipeline

This notebook takes the train and test data and does the following:

- Data processing (cleaning, feature engineering, scaling etc.) to prepare it for model training
- Trains a simple model
- Saves any artifacts that will be used during inference to disk

> This process will likely be captured in a Pipeline asset on Highwind

# Config
RANDOM_SEED = 42
ARTIFACT_SAVE_DIR = "../saved_model/"
TRAIN_DATA_PATH = "../data/train.csv"
TEST_DATA_PATH = "../data/test.csv"
TARGET_COLUMN = "target"
MODEL_ARGS = {
    "C": 1,
    "solver": "lbfgs",
    "fit_intercept": True,
    "random_state": 42,
}
PUSH_TO_HF = True # Whether to push to Hugging Face Hub or not
HF_REPO_NAME = "MelioAI/iris-classifier" # For pushing model to Hugging Face Hub 

import os
import sklearn
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import classification_report
import pandas as pd
import joblib
from skops import card, hub_utils
from tempfile import mkdtemp, mkstemp
from pathlib import Path

## Load data

train_df = pd.read_csv(TRAIN_DATA_PATH)
train_df.head(3)

# Separate features and labels
X_train = train_df.copy()
y_train = X_train.pop(TARGET_COLUMN)

# Check shapes
print(f"X_train: {X_train.shape}")
print(f"y_train: {y_train.shape}")

## Process data

### Optional steps

Insert any optional data processing steps here

# (Optional) Add data cleaning here

# (Optional) Add feature engineering here

### Feature scaling

# Initialise scaler and scale train features
scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)

# Save the scaler for later use
save_scaler_path = os.path.join(ARTIFACT_SAVE_DIR, "scaler.joblib")
joblib.dump(scaler, save_scaler_path)

## Train model

# Define model
model = LogisticRegression(**MODEL_ARGS)

# Train model
model.fit(X_train_scaled, y_train)

# Save the model for later use
save_model_path = os.path.join(ARTIFACT_SAVE_DIR, "model.joblib")
joblib.dump(model, save_model_path)

## (Optional) Save to Hugging Face Hub
Save trained model files to the Hugging Face Hub so that they can be downloaded later. In this step, we use the useful helper functions provided by the [`skops`](https://github.com/skops-dev/skops/tree/main) package.

If `PUSH_TO_HF` is enabled (see top of this notebook), this section will execute. Remember to log into Hugging Face with the CLI by running: `huggingface-cli login` otherwise this section won't work.

# Make temporary local repo dir
local_repo = Path("../hf-repo") # mkdtemp(prefix="skops-")

if PUSH_TO_HF:

    # Initialise HF repo
    hub_utils.init(
        model=Path(save_model_path),
        requirements=[
            f"scikit-learn=={sklearn.__version__}",
            f"joblib=={joblib.__version__}"
        ],
        dst=local_repo,
        task="tabular-classification",
        data=X_train.head(),
        model_format="pickle"
    )

    # Add feature scaler to repo
    hub_utils.add_files(save_scaler_path, dst=local_repo)

if PUSH_TO_HF:

    # Create and populate basic model card
    model_card = card.Card(model=model)
    metadata = card.metadata_from_config(local_repo / "config.json")
    
    # Add model card detail
    limitations = (
        "This model is made for the purposes of showing how to use Highwind only."
    )
    model_description = (
        "This is a Logistic Regression model trained on iris dataset. This model could be"
        " used to predict type of iris flower, given certain dimensions. This model is very basic and"
        " should only be used as an example of how to use Highwind."
    )
    model_card_authors = "MelioAI, ruanmelio"
    usage_code = """
```python
import joblib
from huggingface_hub import hf_hub_download

# Feature scaler
hf_hub_download("MelioAI/iris-classifier", "scaler.joblib")
scaler = joblib.load("scaler.joblib")

# Classifier model
hf_hub_download("MelioAI/iris-classifier", "model.joblib")
model = joblib.load("model.joblib")
```
"""
    model_card.add(
        folded=False,
        **{
            "Model Card Authors": model_card_authors,
            "Intended uses & limitations": limitations,
            "Model description": model_description,
            "Model description/Intended uses & limitations": limitations,
            "How to Get Started with the Model": usage_code
        },
    )

    # Add tags
    model_card.metadata.library_name = "sklearn"
    model_card.metadata.tags = ["sklearn", "tabular-classification"]

    # Save model card
    model_card.save(local_repo / "README.md")

# Remember to log into HF with the CLI by running: huggingface-cli login
if PUSH_TO_HF:

    # Push to HF Hub
    hub_utils.push(
        repo_id=HF_REPO_NAME,
        source=local_repo
    )

## (Optional) model evaluation

Using test set

# Remove scaler and model from memory to prove loading from disk works
del model
del scaler

# Load test data
test_df = pd.read_csv(TEST_DATA_PATH)
print(f"test_df: {test_df.shape}")

# Separate features and labels
X_test = test_df.copy()
y_test = X_test.pop(TARGET_COLUMN)

# Run through same preprocessing steps
# Feature scaling
scaler = joblib.load(os.path.join(ARTIFACT_SAVE_DIR, "scaler.joblib"))
X_test_scaled = scaler.transform(X_test)

# Load model
model = joblib.load(os.path.join(ARTIFACT_SAVE_DIR, "model.joblib"))

# Make predictions on test set
y_pred = model.predict(X_test_scaled)

# Model evaluation report
report = classification_report(y_test, y_pred)
print(report)
